requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0

# LLM clients
openai>=1.50.0
google-genai>=0.8.0

# Configuration and utilities
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Typing and async support
typing-extensions>=4.5.0
numpy>=1.21.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Logging and monitoring
structlog>=23.0.0
